chore(personhog): extract the generic gRPC serving layer into common-grpc
The metrics, load-shedding, connection-tracking and caller task-local
layers in personhog-common were usable only by personhog services,
because depending on them also pulled in personhog-proto, sqlx and the
persons and partitioning modules. They now live in common-grpc, next to
common-metrics and common-health.

personhog-common keeps the semantic refusal helpers, which encode the
router's retry contract, and ClientInFlightGuard, whose metric name is
personhog-specific. GZIP_OVERHEAD_HEADER moves next to the AsyncGzipLayer
that sets it. No metric is renamed.
