+++ rust/Cargo.toml
@@ -19,6 +19,7 @@ members = [
   "common/cookieless",
   "common/database",
   "common/dns",
+  "common/grpc",
   "common/health",
   "common/kafka",
   "common/liveness",
@@ -245,6 +246,7 @@ redis = { version = "0.32.7", features = [
   "tls-rustls-webpki-roots",
   "tokio-rustls-comp",
 ] }
+pin-project = "1.1"
 regex = "1.11.1"
 fancy-regex = "0.17"
 lz-str = "0.2.1"
