+++ rust/common/grpc/Cargo.toml
@@ -1,32 +1,20 @@
 [package]
-name = "personhog-common"
+name = "common-grpc"
 version = "0.1.0"
 edition = "2021"
 publish = false
 
 [dependencies]
-common-grpc = { path = "../common/grpc" }
-personhog-proto = { path = "../personhog-proto" }
-
-bytes = { workspace = true }
-chrono = { workspace = true }
-flate2 = { workspace = true }
 futures = { workspace = true }
 http = { workspace = true }
-http-body = { workspace = true }
-http-body-util = { workspace = true }
 metrics = { workspace = true }
-serde_json = { workspace = true }
-sqlx = { workspace = true }
+pin-project = { workspace = true }
 tokio = { workspace = true }
 tonic = { workspace = true }
 tower = { workspace = true }
-thiserror = { workspace = true }
 tracing = { workspace = true }
-uuid = { workspace = true }
 
 [dev-dependencies]
-rstest = "0.26"
 tokio = { workspace = true, features = ["macros", "rt"] }
 tower = { workspace = true, features = ["util"] }
