+++ rust/personhog-identity/Cargo.toml
@@ -9,6 +9,7 @@ personhog-common = { path = "../personhog-common" }
 personhog-proto = { path = "../personhog-proto" }
 common-database = { path = "../common/database" }
 common-alloc = { path = "../common/alloc" }
+common-grpc = { path = "../common/grpc" }
 common-metrics = { path = "../common/metrics" }
 lifecycle = { path = "../common/lifecycle" }
